requirements.txt
# General modules:
django >3.2,<3.3
python-decouple
Pillow
djangorestframework
django-summernote
django-simple-captcha
crispy-bootstrap5

# Third party modules:
markdown       
django-filter  
drf-yasg[validation]

# Deployment modules:
djangorestframework-simplejwt
gunicorn
django-robots
django_compressor

# email
django-mail-templated

# reformatting
black
flake8
pytest-django
Faker

# background proccess and cache
celery
redis
django-celery-beat
django-redis

# database
psycopg2-binary